Added chat support for Arconaitv
A minor fix and you can now open the chat for Arconaitv.

And on that note, I would like to address Arconaitv directly.

I know it's wildly presumptuous to assume this tiny, little, bash-only project can even appear on your radar, let alone be worth the effort to change the code on your servers. But in the off chance that's what's happening, I feel it's important to open a dialog.

Now I'm not going to beat around the bush. You and I both know what's going on here. And I'm not judging. Servers cost money. Electricity costs money. Food costs money. It doesn't matter what content is being supplied, it's not free to supply it. That's one reason why I'm willing to compromise.

To put some perspective on this, I was able to undo all of your escape sequences with one line of code, as anybody who knows their way around bash could. It only took a few minutes to fix, test, and implement on a global level. It's honestly taking me much longer to write this short letter, which is why I'm only just now getting around to it. The fix was ready for the public weeks ago, but I'm not sharing the code yet because I want to give you some time to consider my proposal.

I want to believe there is a way for our projects to co-exist. I believe you can still rake in ad revenue and, with chat support for your site now having been implemented, all it would take is to show ads in the chat's popout window. That way what few people who may be using this script can take part in your chats, and you can still rake in the ad revenue. It's not a great solution, I know, but neither is a coding war.

And that's what I'm trying to prevent. I'm sure you don't have time for that and frankly neither do I. I don't want to have to keep checking back to see if this still works with your site, and I know you don't want to keep coming back to smack down this little bash script. There is simply no reason why we should exhaust ourselves like that when we can better invest our time with more meaningful pursuits.

So let's save ourselves the time, skip the coding war, and get right to the negotiations. In a few weeks I'm going to release the fix. I feel that will give you enough time to implement ads for the chat's popout. That way you can still get yours and the viewers can still get theirs. It's win-win.

Of course that doesn't have to be the only solution. I'm open to suggestion. If you have other ideas or need more time, let me know. I think you'll find I'm quite easy to work with.