Lots of changes
Can't remember them all but here's a few.

A healthy amount of bug fixes.
LSLGUI will now automatically dig for media links if there is no plugin available for a given link.
Arconai is no longer supported as there is a plugin for it now. Opening the chat is still supported.
It's much easier to add support for new media file types now.
Lots of little changes.