Vaughnlive quick fix and "open link in browser"
I whipped up a quick fix for vaughnlive a while back but I didn't bother updating it here because I've been expecting the plugin to get fixed any day now. But it's been all quiet on the front for over week now so it's hard to say if/when it will get fixed.

So instead, here's my fix. It's not bullet-proof and may not work with every stream, but it's working for everything I want to watch.

Also you can now open links in your web-browser. This is useful for saved links. Especially if you need to troubleshoot something. When the player is closed, it will be in the list of options.