Userspace plugin directory added
You can now have a secondary plugin directory. It's at <your LiveStreamLinkGUI config directory>/plugins. Plugins in this secondary directory take priority over the default plugins. Some websites like being a pain and frequently stop working with Livestreamer/Streamlink. And if you were to find a fix for that site's plugin and wanted to change it manually, you would have to elevate yourself to superuser status, edit the file, and then save. Then you would have to do this with every system that uses Livestreamer/Streamlink. It can be very annoying when all you want to do is turn on a show while you work. But now, you can save plugins you edit yourself to <your LiveStreamLinkGUI config directory>/plugins without needing an admin password, set a cloud service to sync the directory across all of your systems, and they will automatically update when you edit a plugin! So when a fix is found for a website, you only need to make one change on one system and the rest will update, too!