Radiolists were optional this whole time??
A long procrastinated update. A whole year since the last update!

A lot of LSLGUI was rewritten. It's a bit neater now and a little more consistent overall. I don't remember what exactly was updated, but one of the most noteworthy changes is changing from radiolists to standard lists. I've always had a problem with radiolists and no tutorial I read explained that they were optional. So I'm a lot happier with it now. This is a much better interface.

As always, everything is still compatible with previous versions. Simply download the new version and overwrite the old one. If you changed the default configuration directory, be sure to make that one change near the top of the new script before you run it.

Enjoy!